+++ tests/autofix_files/async100_asyncio.py
@@ -2,6 +2,7 @@
 # ANYIO_NO_ERROR
 # BASE_LIBRARY asyncio
 
+# timeout[_at] re-exported in the main asyncio namespace in py3.11
 # mypy: disable-error-code=attr-defined
 # AUTOFIX
 
@@ -10,16 +11,12 @@ import asyncio.timeouts
 
 
 async def foo():
-    # py>=3.11 re-exports these in the main asyncio namespace
     # error: 9, "asyncio", "timeout_at"
     ...
     # error: 9, "asyncio", "timeout"
     ...
 
-    # TODO
-    with asyncio.timeouts.timeout_at(10):
-        ...
-    with asyncio.timeouts.timeout_at(10):
-        ...
-    with asyncio.timeouts.timeout(10):
-        ...
+    # error: 9, "asyncio.timeouts", "timeout_at"
+    ...
+    # error: 9, "asyncio.timeouts", "timeout"
+    ...