+++ tests/autofix_files/async91x_autofix.py
@@ -11,7 +11,6 @@ So we make sure that import is added after it.
 # ARG --enable=ASYNC910,ASYNC911
 
 from typing import Any
-
 import trio
 
 
@@ -135,11 +134,13 @@ async def livelocks():
         ...
 
 
-# this will autofix 910 by adding a checkpoint outside the loop
+# this will autofix 910 by adding a checkpoint outside the loop, which doesn't actually
+# help, and the method still isn't guaranteed to checkpoint in case bar() always returns
+# True.
 async def no_checkpoint():  # ASYNC910: 0, "exit", Statement("function definition", lineno)
     while bar():
         try:
-            await trio.sleep("1")  # type: ignore[arg-type]
-        except ValueError:
+            await foo("1")  # type: ignore[call-arg]
+        except TypeError:
             ...
     await trio.lowlevel.checkpoint()