+++ tests/eval_files/async912_asyncio.py
@@ -6,7 +6,7 @@
 # ASYNC100 supports autofix, but ASYNC912 doesn't, so we must run with NOAUTOFIX
 # NOAUTOFIX
 
-# timeout[_at] added in py3.11
+# timeout[_at] re-exported in the main asyncio namespace in py3.11
 # mypy: disable-error-code=attr-defined
 
 import asyncio
@@ -17,19 +17,37 @@ def bar() -> bool:
 
 
 async def foo():
+    # async100
     async with asyncio.timeout(10):  # ASYNC100: 15, "asyncio", "timeout"
         ...
     async with asyncio.timeout_at(10):  # ASYNC100: 15, "asyncio", "timeout_at"
         ...
+    async with asyncio.timeouts.timeout(
+        10
+    ):  # ASYNC100: 15, "asyncio.timeouts", "timeout"
+        ...
+    async with asyncio.timeouts.timeout_at(
+        10
+    ):  # ASYNC100: 15, "asyncio.timeouts", "timeout_at"
+        ...
 
+    # no errors
     async with asyncio.timeout(10):
         await foo()
     async with asyncio.timeout_at(10):
         await foo()
 
-    async with asyncio.timeout_at(10):  # ASYNC912: 4
+    # async912
+    async with asyncio.timeout_at(10):  # ASYNC912: 15
         if bar():
             await foo()
-    async with asyncio.timeout(10):  # ASYNC912: 4
+    async with asyncio.timeout(10):  # ASYNC912: 15
+        if bar():
+            await foo()
+
+    async with asyncio.timeouts.timeout(10):  # ASYNC912: 15
+        if bar():
+            await foo()
+    async with asyncio.timeouts.timeout_at(10):  # ASYNC912: 15
         if bar():
             await foo()